Comment out variables in default raven-coredump.conf
This means variables will only be set if the user wants that, which
means we can change the defaults in raven-coredump and expect them to
apply unless where users have decided they want something else.

The configuration file is json which doesn't have comments, so I've
just put a #-mark as the first letter in each key.